adding some tests

+++ src/main/scala/com/rydgel/scalagram/responses/Location.scala
@@ -2,7 +2,7 @@ package com.rydgel.scalagram.responses
 
 import play.api.libs.json.Json
 
-case class Location(id: Option[Int], latitude: Option[Float], longitude: Option[Float], name: Option[String])
+case class Location(id: Option[Int], latitude: Option[Double], longitude: Option[Double], name: Option[String])
 object Location {
   implicit val LocationReads = Json.reads[Location]
   implicit val LocationWrites = Json.writes[Location]
